# crates/sbol-db-server/src/v2/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SynBioHub v2 API",
    "version": "0.1.2",
    "description": "The SynBioHub v2 API: SBOL DB's idiomatic, RESTful presentation of the SynBioHub product surface, mounted under `/api/v2`.\n\nSynBioHub v2 calls the same application services as the SynBioHub v1 compatibility adapter: both expose the same product domain and read and write the same dataset, while retaining separate wire contracts. V2 uses proper HTTP verbs (POST creates, PATCH edits, DELETE deletes, and no GET mutates), JSON request bodies, real pagination with a total, `Accept`-driven content negotiation, a single consistent error envelope, and bearer or same-origin browser-session authentication. V1 preserves classic SynBioHub behavior where compatibility requires it.\n\nWhen public browsing is enabled, a missing or unrecognized resource credential is anonymous and scoped to the public graph. An instance with `requireLogin` rejects anonymous resource requests with `401` while retaining public bootstrap and session routes. A mutating verb attempted anonymously is `403`; invalid login credentials are `401`. Reads are ACL-scoped to the caller's authorized graphs, and an out-of-scope object is a non-disclosing `404`.",
    "contact": {
      "name": "SBOL DB",
      "url": "https://github.com/marpaia/sbol-db"
    },
    "license": {
      "name": "MIT OR Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "/api/v2",
...
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "An `Authorization: Bearer <token>` credential. A missing or unrecognized token is anonymous when public browsing is enabled; `requireLogin` instances reject anonymous resource requests."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sbol-db-token",
        "description": "The HttpOnly browser session established by `POST /session`. Bearer credentials take precedence. Unsafe cookie-authenticated requests must be same-origin."
      }
    },
    "schemas": {
...
      "AdminInstance": {
        "type": "object",
        "required": ["name", "instance_url", "uri_prefix", "front_page_text", "allow_public_signup", "require_login", "setup_required"],
        "properties": {
          "name": { "type": "string" },
          "instance_url": { "type": "string" },
          "uri_prefix": { "type": "string" },
          "front_page_text": { "type": "string" },
          "allow_public_signup": { "type": "boolean" },
          "require_login": { "type": "boolean" },
          "setup_required": { "type": "boolean", "readOnly": true }
        }
      },
      "AdminUser": {
        "allOf": [
          { "$ref": "#/components/schemas/SessionUser" }
        ],
        "description": "A safe account projection. Password hashes, reset links, and tokens are never present."
      },
      "AdminUsersResponse": {
        "type": "object",
        "required": ["total", "items"],
        "properties": {
          "total": { "type": "integer" },
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/AdminUser" } }
        }
      },
...
      "AdminAuditResponse": {
        "type": "object",
        "required": ["total", "items"],
        "properties": {
          "total": { "type": "integer" },
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/AdminAuditEvent" } }
        }
      },
      "Confirmation": {
        "type": "object",
        "additionalProperties": false,
        "required": ["confirmation"],
        "properties": {
          "confirmation": { "type": "string", "description": "The exact target-specific phrase documented by the operation." }
        }
      },
      "CompleteBackupTrigger": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "trigger": { "type": "string", "enum": ["manual", "pre_deploy"], "default": "manual" },
          "idempotency_key": { "type": "string", "maxLength": 200, "description": "Required for pre_deploy; use a release or commit identifier." }
        }
      },
      "CompleteBackupStatus": {
        "type": "object",
        "required": ["enabled", "strategy", "components", "recent"],
        "properties": {
          "enabled": { "type": "boolean" },
          "strategy": { "type": "string", "const": "complete_encrypted_checkpoint" },
          "components": { "type": "array", "description": "Generation roots in the artifact. The RocksDB checkpoint includes accounts and durable runtime configuration.", "items": { "type": "string", "enum": ["rocksdb", "blobs", "search", "acme"] } },
          "recent": { "type": "array", "items": { "type": "object" } }
        }
      },
...
      "EdgeSettingsPatch": {
        "type": "object",
        "additionalProperties": false,
        "description": "Every property is optional; supplied values replace pending settings and apply after restart.",
        "properties": {
          "hostname": { "type": "string" },
          "acme_contact": { "type": "string", "format": "email" },
          "acme_directory_url": { "type": "string", "format": "uri" },
          "http_redirect_enabled": { "type": "boolean" },
          "tls_handshake_timeout_secs": { "type": "integer", "minimum": 1, "maximum": 60 },
          "backup_recovery_recipient": { "type": "string" },
          "backup_repository_url": { "type": "string" },
          "backup_interval_secs": { "type": "integer", "minimum": 900, "maximum": 2592000 },
          "backup_local_retention": { "type": "integer", "minimum": 1, "maximum": 30 },
          "minimum_free_bytes": { "type": "integer", "minimum": 1 }
        }
      },
...
      "EdgeAdminSnapshot": {
        "type": "object",
        "required": ["active", "pending", "restart_required", "runtime", "health", "recovery"],
        "properties": {
          "active": { "$ref": "#/components/schemas/EdgeSettings" },
          "pending": { "$ref": "#/components/schemas/EdgeSettings" },
          "restart_required": { "type": "boolean" },
          "runtime": { "$ref": "#/components/schemas/EdgeRuntimeIdentity" },
          "health": { "$ref": "#/components/schemas/EdgeHealthSnapshot" },
          "recovery": { "$ref": "#/components/schemas/EdgeRecoverySnapshot" }
        }
      },
      "Error": {
        "type": "object",
        "description": "The single V2 error envelope. Every non-2xx response carries this shape.",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "object",
            "required": ["code", "message", "status"],
            "properties": {
              "code": { "type": "string", "description": "A stable machine-readable code, shared with the SBOL DB API." },
              "message": { "type": "string", "description": "A human-readable explanation." },
              "status": { "type": "integer", "description": "The HTTP status code, echoed for clients that only read the body." }
            }
          }
        }
      },
...
      "SearchResponse": {
        "type": "object",
        "required": ["items", "total", "offset", "limit", "sort", "direction"],
        "properties": {
          "items": { "type": "array", "items": { "$ref": "#/components/schemas/SearchHit" } },
          "total": { "type": "integer", "description": "The total number of in-scope matches, independent of the returned window." },
          "offset": { "type": "integer" },
          "limit": { "type": "integer" },
          "sort": { "type": "string", "enum": ["relevance", "name", "created", "modified", "iri"] },
          "direction": { "type": "string", "enum": ["asc", "desc"] }
        }
      },
...
    },
    "responses": {
      "Error": {
        "description": "An error, in the V2 envelope.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      }
    },
...
  },
  "security": [
    { "bearerAuth": [] },
    { "cookieAuth": [] }
  ],
  "paths": {
...
    "/admin": {
      "get": {
        "tags": ["Administration"],
        "summary": "Discover administrator capabilities",
        "description": "Returns the active backend capabilities, immutable search deployment, and the SynBioHub v2 admin sections. Requires an authenticated administrator.",
        "operationId": "v2AdminOverview",
        "responses": {
          "200": { "description": "Administrator capability map.", "content": { "application/json": { "schema": { "type": "object" } } } },
          "401": { "$ref": "#/components/responses/Error" },
          "403": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/admin/instance": {
      "get": {
        "tags": ["Administration"], "summary": "Read editable instance settings", "operationId": "v2AdminInstance",
        "responses": {
          "200": { "description": "The safe editable instance projection.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminInstance" } } } },
          "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }
        }
      },
      "patch": {
        "tags": ["Administration"], "summary": "Update instance identity and public policy", "operationId": "v2AdminUpdateInstance",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "additionalProperties": false, "properties": {
          "name": { "type": "string" }, "instance_url": { "type": "string" }, "uri_prefix": { "type": "string" }, "front_page_text": { "type": "string" }, "allow_public_signup": { "type": "boolean" }, "require_login": { "type": "boolean" }
        } } } } },
        "responses": {
          "200": { "description": "Updated instance settings.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminInstance" } } } },
          "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/admin/users": {
      "get": {
        "tags": ["Administration"], "summary": "List safe account profiles", "operationId": "v2AdminUsers",
        "responses": {
          "200": { "description": "Local accounts without credentials or reset material.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminUsersResponse" } } } },
          "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "tags": ["Administration"], "summary": "Create a local account", "operationId": "v2AdminCreateUser",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "additionalProperties": false, "required": ["username", "name", "email", "password"], "properties": {
          "username": { "type": "string" }, "name": { "type": "string" }, "email": { "type": "string", "format": "email" }, "affiliation": { "type": "string" }, "password": { "type": "string", "format": "password", "writeOnly": true }, "is_admin": { "type": "boolean" }, "is_curator": { "type": "boolean" }, "is_member": { "type": "boolean", "default": true }
        } } } } },
        "responses": {
          "201": { "description": "Created account.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminUser" } } } },
          "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/admin/users/{username}": {
      "parameters": [{ "name": "username", "in": "path", "required": true, "schema": { "type": "string" } }],
      "patch": {
        "tags": ["Administration"], "summary": "Update an account profile and roles", "description": "Username and account graph identity remain immutable. Self-demotion and removal of the final administrator are rejected.", "operationId": "v2AdminUpdateUser",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "additionalProperties": false, "properties": {
          "name": { "type": "string" }, "email": { "type": "string", "format": "email" }, "affiliation": { "type": "string" }, "is_admin": { "type": "boolean" }, "is_curator": { "type": "boolean" }, "is_member": { "type": "boolean" }
        } } } } },
        "responses": { "200": { "description": "Updated account.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminUser" } } } }, "400": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } }
      },
      "delete": {
        "tags": ["Administration"], "summary": "Delete an account", "description": "Requires `DELETE {username}` exactly. Self-delete and deletion of the final administrator are rejected and every attempt is audited.", "operationId": "v2AdminDeleteUser",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Confirmation" } } } },
        "responses": { "204": { "description": "Account deleted." }, "400": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/admin/integrations": {
      "get": {
        "tags": ["Administration"], "summary": "Read federation, remote, and plugin settings", "description": "Remote credential fields are recursively redacted.", "operationId": "v2AdminIntegrations",
        "responses": { "200": { "description": "Secret-safe integration settings.", "content": { "application/json": { "schema": { "type": "object" } } } }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/admin/federation": {
      "post": {
        "tags": ["Administration"], "summary": "Join a Web of Registries", "operationId": "v2AdminFederate",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["administrator_email", "url"], "properties": { "administrator_email": { "type": "string", "format": "email" }, "url": { "type": "string", "format": "uri" } } } } } },
        "responses": { "200": { "description": "Membership saved.", "content": { "application/json": { "schema": { "type": "object" } } } }, "400": { "$ref": "#/components/responses/Error" }, "501": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/admin/federation/sync": {
      "post": {
        "tags": ["Administration"], "summary": "Synchronize registry routes", "operationId": "v2AdminFederationSync",
        "responses": { "200": { "description": "Synchronization count.", "content": { "application/json": { "schema": { "type": "object" } } } }, "400": { "$ref": "#/components/responses/Error" }, "501": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/admin/registries": {
      "post": {
        "tags": ["Administration"], "summary": "Upsert a registry route", "operationId": "v2AdminSaveRegistry",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["uri", "url"], "properties": { "uri": { "type": "string" }, "url": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Registry route saved.", "content": { "application/json": { "schema": { "type": "object" } } } }, "400": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/admin/registries/{uri}": {
      "delete": {
        "tags": ["Administration"], "summary": "Delete a registry route", "description": "Requires `DELETE REGISTRY {uri}` exactly.", "operationId": "v2AdminDeleteRegistry",
        "parameters": [{ "name": "uri", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Confirmation" } } } },
        "responses": { "204": { "description": "Registry route deleted." }, "400": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/admin/remotes": {
      "post": {
        "tags": ["Administration"], "summary": "Upsert an ICE or Benchling remote", "description": "The configuration is write-only for secret fields; subsequent reads redact them.", "operationId": "v2AdminSaveRemote",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } },
        "responses": { "200": { "description": "Remote id saved.", "content": { "application/json": { "schema": { "type": "object" } } } }, "400": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/admin/remotes/{id}": {
      "delete": {
        "tags": ["Administration"], "summary": "Delete a remote", "description": "Requires `DELETE REMOTE {id}` exactly.", "operationId": "v2AdminDeleteRemote",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Confirmation" } } } },
        "responses": { "204": { "description": "Remote deleted." }, "400": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/admin/plugins": {
      "post": {
        "tags": ["Administration"], "summary": "Add or replace a plugin", "operationId": "v2AdminSavePlugin",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["category", "id", "name", "url"], "properties": { "category": { "type": "string" }, "id": { "type": "string" }, "name": { "type": "string" }, "url": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Plugin saved.", "content": { "application/json": { "schema": { "type": "object" } } } }, "400": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/admin/plugins/{category}/{id}": {
      "delete": {
        "tags": ["Administration"], "summary": "Delete a plugin", "description": "Requires `DELETE PLUGIN {category}/{id}` exactly.", "operationId": "v2AdminDeletePlugin",
        "parameters": [{ "name": "category", "in": "path", "required": true, "schema": { "type": "string" } }, { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Confirmation" } } } },
        "responses": { "204": { "description": "Plugin deleted." }, "400": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/admin/jobs": {
      "get": {
        "tags": ["Administration"], "summary": "List jobs", "operationId": "v2AdminJobs",
        "parameters": [{ "name": "kind", "in": "query", "schema": { "type": "string" } }, { "name": "status", "in": "query", "schema": { "type": "string" } }, { "name": "queue", "in": "query", "schema": { "type": "string" } }, { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100 } }],
        "responses": { "200": { "description": "Matching jobs.", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object" } } } } }, "400": { "$ref": "#/components/responses/Error" } }
      },
      "post": {
        "tags": ["Administration"], "summary": "Enqueue a job", "operationId": "v2AdminEnqueueJob",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["kind", "payload"], "properties": { "kind": { "type": "string" }, "payload": {}, "queue": { "type": "string" }, "priority": { "type": "integer" }, "max_attempts": { "type": "integer" }, "idempotency_key": { "type": "string" }, "correlation_id": { "type": "string", "format": "uuid" } } } } } },
        "responses": { "200": { "description": "Enqueue result.", "content": { "application/json": { "schema": { "type": "object" } } } }, "400": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/admin/jobs/{id}": {
      "get": {
        "tags": ["Administration"], "summary": "Read one job", "operationId": "v2AdminJob",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Job record.", "content": { "application/json": { "schema": { "type": "object" } } } }, "400": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/admin/jobs/{id}/attempts": {
      "get": {
        "tags": ["Administration"], "summary": "List job attempts", "operationId": "v2AdminJobAttempts",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": { "200": { "description": "Attempt records.", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object" } } } } }, "404": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/admin/jobs/{id}/logs": {
      "get": {
        "tags": ["Administration"], "summary": "List job log records", "operationId": "v2AdminJobLogs",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }, { "name": "after_id", "in": "query", "schema": { "type": "integer" } }, { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 500 } }],
        "responses": { "200": { "description": "Job log records.", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object" } } } } }, "404": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/admin/jobs/{id}/cancel": {
      "post": {
        "tags": ["Administration"], "summary": "Cancel a queued or running job", "description": "Requires `CANCEL JOB {id}` exactly.", "operationId": "v2AdminCancelJob",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Confirmation" } } } },
        "responses": { "200": { "description": "Whether the job was cancelled.", "content": { "application/json": { "schema": { "type": "object" } } } }, "400": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/admin/ontologies": {
      "get": {
        "tags": ["Administration"], "summary": "List loaded ontologies", "operationId": "v2AdminOntologies",
        "responses": { "200": { "description": "Loaded ontologies.", "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object" } } } } } }
      },
      "post": {
        "tags": ["Administration"], "summary": "Load an ontology", "operationId": "v2AdminLoadOntology",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["prefix"], "properties": { "prefix": { "type": "string" }, "url": { "type": "string" }, "name": { "type": "string" } } } } } },
        "responses": { "200": { "description": "Ontology import report.", "content": { "application/json": { "schema": { "type": "object" } } } }, "400": { "$ref": "#/components/responses/Error" } }
      }
    },
...
    "/admin/backup": {
      "get": {
        "tags": ["Administration"], "summary": "Read complete-backup capability and recent jobs", "operationId": "v2AdminBackupStatus",
        "responses": { "200": { "description": "Complete-backup status.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CompleteBackupStatus" } } } } }
      },
      "post": {
        "tags": ["Administration"], "summary": "Enqueue a complete encrypted backup", "description": "Manual UI and pre-deploy requests enqueue the same durable complete_backup job used by the scheduler. The artifact includes RocksDB data, accounts and durable configuration, blobs, search state, and ACME state.", "operationId": "v2AdminTriggerBackup",
        "requestBody": { "required": false, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CompleteBackupTrigger" } } } },
        "responses": { "202": { "description": "Complete-backup job accepted.", "content": { "application/json": { "schema": { "type": "object" } } } }, "400": { "$ref": "#/components/responses/Error" }, "501": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/admin/edge": {
      "get": {
        "tags": ["Administration"], "summary": "Read active and pending edge settings and live health", "operationId": "v2AdminEdgeStatus",
        "responses": { "200": { "description": "Edge runtime configuration and health.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EdgeAdminSnapshot" } } } }, "501": { "$ref": "#/components/responses/Error" } }
      },
      "patch": {
        "tags": ["Administration"], "summary": "Validate and persist pending edge settings", "description": "Settings are stored in the backend and included in complete RocksDB checkpoints. The running process remains on the active settings until restart. Credentials and recovery identities are not accepted.", "operationId": "v2AdminUpdateEdgeSettings",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EdgeSettingsPatch" } } } },
        "responses": { "200": { "description": "Updated pending settings and restart state.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EdgeAdminSnapshot" } } } }, "400": { "$ref": "#/components/responses/Error" }, "501": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/admin/audit": {
      "get": {
        "tags": ["Administration"], "summary": "Read append-only administrator activity", "operationId": "v2AdminAudit",
        "parameters": [{ "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 1000, "default": 100 } }],
        "responses": { "200": { "description": "Newest events first.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AdminAuditResponse" } } } } }
      }
    },
    "/search": {
      "get": {
        "tags": ["Search"],
        "summary": "Discover registry objects",
        "description": "Normalized ACL-scoped discovery with text, biological facets, collection/owner/provenance/date narrowing, deterministic sorting, exact totals, and stable paging. The wire uses query parameters rather than V1's path grammar.",
        "operationId": "v2Search",
        "parameters": [
          { "name": "q", "in": "query", "required": false, "description": "The free-text term. Absent browses the whole in-scope corpus.", "schema": { "type": "string" } },
          { "name": "type", "in": "query", "required": false, "description": "Restrict to one rdf:type, given as a full IRI.", "schema": { "type": "string" } },
          { "name": "role", "in": "query", "required": false, "description": "Restrict to one SBOL role, given as a full IRI.", "schema": { "type": "string" } },
          { "name": "collection", "in": "query", "required": false, "description": "Restrict to direct members of this collection IRI.", "schema": { "type": "string" } },
          { "name": "owner", "in": "query", "required": false, "description": "Restrict to objects owned by this user graph IRI.", "schema": { "type": "string" } },
          { "name": "provenance", "in": "query", "required": false, "description": "Case-insensitive substring over mutable provenance.", "schema": { "type": "string" } },
          { "name": "created_after", "in": "query", "required": false, "schema": { "type": "string", "format": "date" } },
          { "name": "created_before", "in": "query", "required": false, "schema": { "type": "string", "format": "date" } },
          { "name": "modified_after", "in": "query", "required": false, "schema": { "type": "string", "format": "date" } },
          { "name": "modified_before", "in": "query", "required": false, "schema": { "type": "string", "format": "date" } },
          { "name": "sort", "in": "query", "required": false, "schema": { "type": "string", "enum": ["relevance", "name", "created", "modified", "iri"], "default": "relevance" } },
          { "name": "direction", "in": "query", "required": false, "description": "Defaults to ascending for name/IRI and descending otherwise.", "schema": { "type": "string", "enum": ["asc", "desc"] } },
          { "name": "offset", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 0, "default": 0 } },
          { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 1000, "default": 50 } }
        ],
        "responses": {
          "200": {
            "description": "A deterministic page of discovery hits with an exact total.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/SearchResponse" } }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "401": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "tags": ["Search"],
        "summary": "Execute a pluggable search strategy",
        "description": "Executes a structured request against a registered strategy under the caller's ACL graph scope. This is additive to the compatibility GET operation.",
        "operationId": "v2StructuredSearch",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/StructuredSearchRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "A strategy-labelled page whose total explicitly declares its guarantee.",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/StructuredSearchPage" } }
            }
          },
          "400": { "$ref": "#/components/responses/Error" },
          "503": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  }
}